refactor(harness-runtime): separate profile admission and binding
